Split autoTestERC20
