+++ contracts/test/TestERC20.sol
@@ -32,74 +32,3 @@ contract TestERC20 is ERC20 {
         _burn(msg.sender, amount);
     }
 }
-
-contract TestERC20V2 is ERC20 {
-    uint8 private _dec;
-    mapping(address => uint256) private _balances;
-    mapping(address => bool) public touched;
-    uint256 private _default;
-
-    constructor(
-        string memory name,
-        string memory symbol,
-        uint8 dec_,
-        uint256 default_
-    ) ERC20(name, symbol) {
-        _dec = dec_;
-        _default = default_;
-    }
-
-    function decimals() public view virtual override returns (uint8) {
-        return _dec;
-    }
-
-    function mint(uint256 value) public {
-        touched[msg.sender] = true;
-        _balances[msg.sender] = value;
-        emit Transfer(address(0), msg.sender, value);
-    }
-
-    function balanceOf(address account)
-        public
-        view
-        virtual
-        override
-        returns (uint256)
-    {
-        if (touched[account]) {
-            return _balances[account];
-        } else {
-            return _default;
-        }
-    }
-
-    function _transfer(
-        address from,
-        address to,
-        uint256 amount
-    ) internal virtual override {
-        require(from != address(0), "TF F 0 Addr");
-        require(to != address(0), "TF T 0 Addr");
-
-        _beforeTokenTransfer(from, to, amount);
-        if (from != address(0) && !touched[from]) {
-            touched[from] = true;
-            _balances[from] = _default;
-        }
-
-        if (!touched[to]) {
-            touched[to] = true;
-        }
-
-        uint256 fromBalance = _balances[from];
-        require(fromBalance >= amount, "TF OVER BALANCE");
-        unchecked {
-            _balances[from] = fromBalance - amount;
-        }
-        _balances[to] += amount;
-
-        emit Transfer(from, to, amount);
-
-        _afterTokenTransfer(from, to, amount);
-    }
-}
